+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - REST API: jwt authentication based on JWT scopes array
 - REST API: POST `/repos/export` creates a zip for the specified repository
 - REST API: GET `/repos/export` to download archived repo
+- Added `filter.latest` parameter to REST API `/inventory` listing
 - generating `manufacturers.txt` and `mpns.txt` for improving static hosting (e.g. github, github pages)
 
 ### Fixed